fastapi_service/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
SQLAlchemy==1.4.41
aiosqlite==0.17.0
python-jose==3.3.0
passlib==1.7.4
databases==0.6.2
python-multipart==0.0.5